feat: fix login sequence so player can join the world
Restructure the packet sequence to match Dragonfly/gophertunnel protocol:
- RequestChunkRadius now responds with ChunkRadiusUpdated + BiomeDefinitionList
  (empty) + PlayerSpawn + CreativeContent (empty), then waits for client to
  send SetLocalPlayerAsInitialized before streaming any chunk data
- Add on_enter_play system that sends NetworkChunkPublisher + LevelChunkPackets
  (LIMITED sub-chunk mode) when entering Play state
- Remove chunk sending from the Setup state handler entirely

Previously PlayerSpawn was sent too early (before RequestChunkRadius) and
chunks were all sent at once without the client being in loading state,
causing the client to never request sub-chunks.
